make THKSegment::CheckType a signed integer

+++ mhw-thk-extender/thk_extender.cpp
@@ -10,7 +10,7 @@ THKFunction_t THKExtender::s_Target = (THKFunction_t)0x141343650;
 
 uint64_t THKExtender::AiCheckHook(void* cThinkEm, THKSegment* segment, void* cThinkMgr)
 {
-	int32_t check = static_cast<int32_t>(segment->CheckType);
+	int32_t check = segment->CheckType;
 
 	if (check < 0)
 		segment->CheckType = -segment->CheckType;
